fix default metadata fail
