+++ features/automation/common/state/automationDefaultMetadata.ts
@@ -1,0 +1,11 @@
+export const defaultMetadata = {
+  debtToken: '',
+  positionLabel: '',
+  ratioLabel: '',
+  validation: {
+    creationErrors: [],
+    creationWarnings: [],
+    cancelErrors: [],
+    cancelWarnings: [],
+  },
+}
