+++ features/automation/protection/autoSell/state/automationAutoSellMetadata.ts
@@ -1,8 +1,6 @@
 import { AutoBSFormChange } from 'features/automation/common/state/autoBSFormChange'
-import {
-  AutomationCommonMetadata,
-  defaultMetadata,
-} from 'features/automation/common/state/automationMetadata'
+import { defaultMetadata } from 'features/automation/common/state/automationDefaultMetadata'
+import { AutomationCommonMetadata } from 'features/automation/common/state/automationMetadata'
 
 export interface AutomationAutoSellMetadata extends AutomationCommonMetadata<AutoBSFormChange> {}
